Rename from main to index
This will rename from main to index.
