+++ scripts/boj/index.js
