+++ scripts/solvedAC/index.js
